Struct language_e2e_tests::account::Account

pub struct Account {
    pub privkey: Ed25519PrivateKey,
    pub pubkey: Ed25519PublicKey,
    /* private fields */
}

Details about a Diem account.

Tests will typically create a set of Account instances to run transactions on. This type encodes the logic to operate on and verify operations on any Diem account.

Fields
privkey: Ed25519PrivateKey

The current private key for this account.

pubkey: Ed25519PublicKey

The current public key for this account.

Implementations

impl Account

pub fn new() -> Self

Creates a new account in memory.

The account returned by this constructor is a purely logical entity, meaning that it does not automatically get added to the Diem store. To add an account to the store, use AccountData instances with FakeExecutor::add_account_data. This function returns distinct values upon every call.

pub fn new_from_seed(seed: &mut KeyGen) -> Self

Creates a new account in memory given a random seed.

pub fn with_keypair(privkey: Ed25519PrivateKey, pubkey: Ed25519PublicKey) -> Self

Creates a new account with the given keypair.

Like with Account::new, the account returned by this constructor is a purely logical entity.

pub fn new_validator(
addr: AccountAddress,
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey
) -> Self

Creates a new account with the given addr and key pair

Like with Account::new, the account returned by this constructor is a purely logical entity.

pub fn new_genesis_account(address: AccountAddress) -> Self

Creates a new account in memory representing an account created in the genesis transaction.

The address will be [address], which should be an address for a genesis account and the account will use [GENESIS_KEYPAIR][struct@GENESIS_KEYPAIR] as its keypair.

pub fn new_diem_root() -> Self

Creates a new account representing the diem root account in memory.

The address will be diem_root_address, and the account will use [GENESIS_KEYPAIR][struct@GENESIS_KEYPAIR] as its keypair.

pub fn new_testing_dd() -> Self

pub fn new_blessed_tc() -> Self

Creates a new account representing treasury compliance in memory. The account will use [GENESIS_KEYPAIR][struct@GENESIS_KEYPAIR] as its keypair.

pub fn address(&self) -> &AccountAddress

Returns the address of the account. This is a hash of the public key the account was created with.

The address does not change if the account’s keys are rotated.

pub fn make_account_access_path(&self) -> AccessPath

Returns the AccessPath that describes the Account resource instance.

Use this to retrieve or publish the Account blob.

pub fn make_event_generator_access_path(&self) -> AccessPath

Returns the AccessPath that describes the EventHandleGenerator resource instance.

Use this to retrieve or publish the EventHandleGenerator blob.

pub fn make_balance_access_path(
&self,
balance_currency_code: Identifier
) -> AccessPath

Returns the AccessPath that describes the Account balance resource instance.

Use this to retrieve or publish the Account balance blob.

pub fn make_access_path(&self, tag: StructTag) -> AccessPath

pub fn rotate_key(&mut self, privkey: Ed25519PrivateKey, pubkey: Ed25519PublicKey)

Changes the keys for this account to the provided ones.

pub fn auth_key(&self) -> Vec<u8>

Computes the authentication key for this account, as stored on the chain.

This is the same as the account’s address if the keys have never been rotated.

pub fn auth_key_prefix(&self) -> Vec<u8>

Return the first 16 bytes of the account’s auth key

pub fn transaction(&self) -> TransactionBuilder