Struct language_e2e_tests::account::AccountData

pub struct AccountData { /* fields omitted */ }

Represents an account along with initial state about it.

AccountData captures the initial state needed to create accounts for tests.

Methods

impl AccountData

pub fn new(balance: u64, sequence_number: u64) -> Self

Creates a new AccountData with a new account.

Most tests will want to use this constructor.

pub fn with_account(
account: Account,
balance: u64,
sequence_number: u64
) -> Self

Creates a new AccountData with the provided account.

pub fn with_keypair(
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey,
balance: u64,
sequence_number: u64
) -> Self

Creates a new AccountData with the provided account.

pub fn with_account_and_event_counts(
account: Account,
balance: u64,
sequence_number: u64,
sent_events_count: u64,
received_events_count: u64,
delegated_key_rotation_capability: bool,
delegated_withdrawal_capability: bool
) -> Self

Creates a new AccountData with custom parameters.

pub fn rotate_key(
&mut self,
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey
)

Changes the keys for this account to the provided ones.

pub fn to_resource(&self) -> Value

Creates and returns a resource [Value] for this data.

pub fn make_access_path(&self) -> AccessPath

Returns the AccessPath that describes the Account resource instance.

Use this to retrieve or publish the Account blob.

pub fn address(&self) -> &AccountAddress

Returns the address of the account. This is a hash of the public key the account was created with.

The address does not change if the account's [keys are rotated][AccountData::rotate_key].

pub fn account(&self) -> &Account

Returns the underlying [Account] instance.

pub fn into_account(self) -> Account

Converts this data into an Account instance.

pub fn balance(&self) -> u64

Returns the initial balance.

pub fn sequence_number(&self) -> u64

Returns the initial sequence number.

pub fn sent_events_key(&self) -> &[u8]

Returns the unique key for this sent events stream.

pub fn sent_events_count(&self) -> u64

Returns the initial sent events count.

pub fn received_events_key(&self) -> &[u8]

Returns the unique key for this received events stream.

pub fn received_events_count(&self) -> u64

Returns the initial received events count.

impl AccountData

pub fn strategy(
balance_strategy: impl Strategy<Value = u64>
) -> impl Strategy<Value = Self>

Returns a [Strategy] that creates AccountData instances.