Function language_e2e_tests::account_universe::txn_one_account_result

pub fn txn_one_account_result(
    sender: &mut AccountCurrent, 
    amount: u64, 
    gas_cost: u64, 
    low_gas_cost: u64
) -> (TransactionStatus, bool)

Computes the result for running a transfer out of one account. Also updates the account to reflect this transaction.

The return value is a pair of the expected status and whether the transaction was successful.