Function language_e2e_tests::compile::compile_module_with_address

pub fn compile_module_with_address(
    address: &AccountAddress, 
    file_name: &str, 
    code: &str
) -> TransactionPayload

Compile the provided Move code into a blob which can be used as the code to be published (a Module).