Struct language_e2e_tests::executor::FakeExecutor

pub struct FakeExecutor { /* private fields */ }

Provides an environment to run a VM instance.

This struct is a mock in-memory implementation of the Diem executor.

Implementations

impl FakeExecutor

pub fn from_genesis(write_set: &WriteSet) -> Self

Creates an executor from a genesis WriteSet.

pub fn from_saved_genesis(saved_genesis_blob: &[u8]) -> Self

Create an executor from a saved genesis blob

pub fn from_genesis_file() -> Self

Creates an executor from the genesis file GENESIS_FILE_LOCATION

pub fn from_fresh_genesis() -> Self

Creates an executor using the standard genesis.

pub fn allowlist_genesis() -> Self

pub fn from_genesis_with_options(publishing_options: VMPublishingOption) -> Self

Creates an executor from the genesis file GENESIS_FILE_LOCATION with script/module publishing options given by publishing_options. These can only be either Open or CustomScript.

pub fn no_genesis() -> Self

Creates an executor in which no genesis state has been applied yet.

pub fn set_golden_file(&mut self, test_name: &str)

pub fn stdlib_only_genesis() -> Self

Creates an executor with only the standard library Move modules published and not other initialization done.

pub fn custom_genesis(
genesis_modules: &[Vec<u8>],
validator_accounts: Option<usize>,
publishing_options: VMPublishingOption
) -> Self

Creates fresh genesis from the stdlib modules passed in.

pub fn create_raw_account(&mut self) -> Account

Create one instance of AccountData without saving it to data store.

pub fn create_raw_account_data(
&mut self,
balance: u64,
seq_num: u64
) -> AccountData

Create one instance of AccountData without saving it to data store.

pub fn create_xdx_raw_account_data(
&mut self,
balance: u64,
seq_num: u64
) -> AccountData

Create one instance of AccountData with XDX balance without saving it to data store.

pub fn create_accounts(
&mut self,
size: usize,
balance: u64,
seq_num: u64
) -> Vec<Account>

Creates a number of Account instances all with the same balance and sequence number, and publishes them to this executor’s data store.

pub fn apply_write_set(&mut self, write_set: &WriteSet)

Applies a WriteSet to this executor’s data store.

pub fn add_account_data(&mut self, account_data: &AccountData)

Adds an account to this executor’s data store.

pub fn add_module(&mut self, module_id: &ModuleId, module_blob: Vec<u8>)

Adds a module to this executor’s data store.

Does not do any sort of verification on the module.

pub fn read_account_resource(&self, account: &Account) -> Option<AccountResource>

Reads the resource [Value] for an account from this executor’s data store.

pub fn read_account_resource_at_address(
&self,
addr: &AccountAddress
) -> Option<AccountResource>

Reads the resource [Value] for an account under the given address from this executor’s data store.

pub fn read_balance_resource(
&self,
account: &Account,
balance_currency_code: Identifier
) -> Option<BalanceResource>

Reads the balance resource value for an account from this executor’s data store with the given balance currency_code.

pub fn read_balance_resource_at_address(
&self,
addr: &AccountAddress,
balance_currency_code: Identifier
) -> Option<BalanceResource>

Reads the balance resource value for an account under the given address from this executor’s data store with the given balance currency_code.

pub fn execute_block(
&self,
txn_block: Vec<SignedTransaction>
) -> Result<Vec<TransactionOutput>, VMStatus>

Executes the given block of transactions.

Typical tests will call this method and check that the output matches what was expected. However, this doesn’t apply the results of successful transactions to the data store.

pub fn execute_block_and_keep_vm_status(
&self,
txn_block: Vec<SignedTransaction>
) -> Result<Vec<(VMStatus, TransactionOutput)>, VMStatus>

Alternate form of ‘execute_block’ that keeps the vm_status before it goes into the TransactionOutput

pub fn execute_and_apply(
&mut self,
transaction: SignedTransaction
) -> TransactionOutput

Executes the transaction as a singleton block and applies the resulting write set to the data store. Panics if execution fails

pub fn execute_transaction_block(
&self,
txn_block: Vec<Transaction>
) -> Result<Vec<TransactionOutput>, VMStatus>

pub fn execute_transaction(&self, txn: SignedTransaction) -> TransactionOutput

pub fn read_from_access_path(&self, path: &AccessPath) -> Option<Vec<u8>>

Get the blob for the associated AccessPath

pub fn verify_transaction(&self, txn: SignedTransaction) -> VMValidatorResult

Verifies the given transaction by running it through the VM verifier.

pub fn get_state_view(&self) -> &FakeDataStore

pub fn new_block(&mut self)

pub fn new_block_with_timestamp(&mut self, time_stamp: u64)

pub fn set_block_time(&mut self, new_block_time: u64)

pub fn get_block_time(&mut self) -> u64

pub fn exec(
&mut self,
module_name: &str,
function_name: &str,
type_params: Vec<TypeTag>,
args: Vec<Vec<u8>>
)

pub fn try_exec(
&mut self,
module_name: &str,
function_name: &str,
type_params: Vec<TypeTag>,
args: Vec<Vec<u8>>
) -> Result<WriteSet, VMStatus>