Module language_e2e_tests::gas_costs

Gas costs for common transactions.

Structs

CREATE_ACCOUNT_FIRST
The gas cost of a first time create-account transaction.

CREATE_ACCOUNT_NEXT
The gas cost of a create-account transaction.

CREATE_ACCOUNT_TOO_LOW_FIRST
The gas cost of a create-account transaction where the sender has an insufficient balance.

CREATE_ACCOUNT_TOO_LOW_NEXT
The gas cost of a create-account transaction where the sender has an insufficient balance.

CREATE_EXISTING_ACCOUNT_FIRST
The gas cost of a create-account transaction where the receiver already exists.

CREATE_EXISTING_ACCOUNT_NEXT
The gas cost of a create-account transaction where the receiver already exists.

PEER_TO_PEER
The gas cost of a peer-to-peer transaction.

PEER_TO_PEER_NEW_RECEIVER_FIRST
The gas cost of a peer-to-peer transaction that creates a new account.

PEER_TO_PEER_NEW_RECEIVER_NEXT
The gas cost of a peer-to-peer transaction that creates a new account.

PEER_TO_PEER_NEW_RECEIVER_TOO_LOW_FIRST
The gas cost of a peer-to-peer transaction that tries to create a new account, but fails because of an insufficient balance.

PEER_TO_PEER_NEW_RECEIVER_TOO_LOW_NEXT
The gas cost of a peer-to-peer transaction that tries to create a new account, but fails because of an insufficient balance.

PEER_TO_PEER_TOO_LOW
The gas cost of a peer-to-peer transaction with an insufficient balance.

ROTATE_KEY
The gas cost of a rotate-key transaction.

Constants

TXN_RESERVED
The gas each transaction is configured to reserve. If the gas available in the account, converted to microlibra, falls below this threshold, transactions are expected to fail with an insufficient balance.