Function language_e2e_tests::versioning::run_with_versions

pub fn run_with_versions<ParamExec, F>(
    test_golden_prefix: &str, 
    versions: &[u64], 
    starting_state: ParamExec, 
    test_func: F
) where
    F: Fn(VersionedTestEnv),
    ParamExec: Fn(u64) -> VersionedTestEnv, 

This is takes a test body parametrized by a VersionedTestEnv, and the versions to test against The starting state of the VersionedTestEnv for each version number is determined by the starting_state function.