Crate language_e2e_tests

Test infrastructure for the Diem VM.

This crate contains helpers for executing tests against the Diem VM.

Modules

account

Test infrastructure for modeling Diem accounts.

account_universe

A model to test properties of common Diem transactions.

common_transactions

Support for encoding transactions for common situations.

compile

Support for compiling scripts and modules in tests.

currencies

data_store

Support for mocking the Diem data store.

execution_strategies

executor

Support for running the VM to execute and verify transactions.

gas_costs

Gas costs for common transactions.

on_chain_configs

utils

versioning

Macros

assert_prologue_disparity

assert_prologue_parity

current_function_name

Returns the name of the current function. This macro is used to derive the name for the golden file of each test case.

test_with_different_versions

Functions

assert_status_eq

transaction_status_eq