Crate language_e2e_tests

Test infrastructure for the Libra VM.

This crate contains helpers for executing tests against the Libra VM.

Modules

account
Test infrastructure for modeling Libra accounts.

account_universe
A model to test properties of common Libra transactions.

common_transactions
Support for encoding transactions for common situations.

compile
Support for compiling scripts and modules in tests.

data_store
Support for mocking the Libra data store.

executor
Support for running the VM to execute and verify transactions.

gas_costs
Gas costs for common transactions.

keygen

Macros

assert_prologue_disparity

assert_prologue_parity

Functions

assert_status_eq

transaction_status_eq