Struct language_e2e_tests::account::Account

pub struct Account {
    pub privkey: Ed25519PrivateKey,
    pub pubkey: Ed25519PublicKey,
    // some fields omitted
}

Details about a Libra account.

Tests will typically create a set of Account instances to run transactions on. This type encodes the logic to operate on and verify operations on any Libra account.

Fields
privkey: Ed25519PrivateKey

The current private key for this account.

pubkey: Ed25519PublicKey

The current public key for this account.

Methods

impl Account

pub fn new() -> Self

Creates a new account in memory.

The account returned by this constructor is a purely logical entity, meaning that it does not automatically get added to the Libra store. To add an account to the store, use AccountData instances with FakeExecutor::add_account_data. This function returns distinct values upon every call.

pub fn with_keypair(
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey
) -> Self

Creates a new account with the given keypair.

Like with Account::new, the account returned by this constructor is a purely logical entity.

pub fn new_genesis_account(address: AccountAddress) -> Self

Creates a new account in memory representing an account created in the genesis transaction.

The address will be [address], which should be an address for a genesis account and the account will use [GENESIS_KEYPAIR][struct@GENESIS_KEYPAIR] as its keypair.

pub fn new_association() -> Self

Creates a new account representing the association in memory.

The address will be association_address, and the account will use [GENESIS_KEYPAIR][struct@GENESIS_KEYPAIR] as its keypair.

pub fn address(&self) -> &AccountAddress

Returns the address of the account. This is a hash of the public key the account was created with.

The address does not change if the account's keys are rotated.

pub fn make_account_access_path(&self) -> AccessPath

Returns the AccessPath that describes the Account resource instance.

Use this to retrieve or publish the Account blob.

pub fn make_event_generator_access_path(&self) -> AccessPath

Returns the AccessPath that describes the EventHandleGenerator resource instance.

Use this to retrieve or publish the EventHandleGenerator blob.

pub fn make_balance_access_path(
&self,
balance_currency_code: Identifier
) -> AccessPath

Returns the AccessPath that describes the Account balance resource instance.

Use this to retrieve or publish the Account balance blob.

pub fn make_account_type_access_path(
&self,
account_specifier: AccountTypeSpecifier
) -> AccessPath

Returns the AccessPath that describes the Account type resource instance.

Use this to retrieve or publish the Account AccountType blob.

pub fn rotate_key(
&mut self,
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey
)

Changes the keys for this account to the provided ones.

pub fn auth_key(&self) -> Vec<u8>

Computes the authentication key for this account, as stored on the chain.

This is the same as the account's address if the keys have never been rotated.

pub fn auth_key_prefix(&self) -> Vec<u8>

Return the first 16 bytes of the account's auth key

pub fn create_user_txn(
&self,
payload: TransactionPayload,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> SignedTransaction

Returns a SignedTransaction with a payload and this account as the sender.

This is the most generic way to create a transaction for testing. Max gas amount and gas unit price are ignored for WriteSet transactions.

pub fn create_raw_user_txn(
address: AccountAddress,
payload: TransactionPayload,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> RawTransaction

pub fn create_signed_txn_with_args(
&self,
program: Vec<u8>,
ty_args: Vec<TypeTag>,
args: Vec<TransactionArgument>,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> SignedTransaction

Returns a SignedTransaction with the arguments defined in args and this account as the sender.

pub fn create_raw_txn_with_args(
address: AccountAddress,
program: Vec<u8>,
ty_args: Vec<TypeTag>,
args: Vec<TransactionArgument>,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> RawTransaction

pub fn create_signed_txn_with_args_and_sender(
&self,
sender: AccountAddress,
program: Vec<u8>,
ty_args: Vec<TypeTag>,
args: Vec<TransactionArgument>,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> SignedTransaction

Returns a SignedTransaction with the arguments defined in args and a custom sender.

The transaction is signed with the key corresponding to this account, not the custom sender.

pub fn create_raw_txn_with_args_and_sender(
sender: AccountAddress,
program: Vec<u8>,
ty_args: Vec<TypeTag>,
args: Vec<TransactionArgument>,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> RawTransaction

pub fn create_signed_txn_impl(
&self,
sender: AccountAddress,
program: TransactionPayload,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> SignedTransaction

Returns a SignedTransaction with the arguments defined in args and a custom sender.

The transaction is signed with the key corresponding to this account, not the custom sender.

pub fn signed_script_txn(
&self,
script: Script,
sequence_number: u64
) -> SignedTransaction

Create a transaction containing script signed by sender with default values for gas cost, gas price, expiration time, and currency type.

pub fn create_raw_txn_impl(
sender: AccountAddress,
program: TransactionPayload,
sequence_number: u64,
max_gas_amount: u64,
gas_unit_price: u64,
gas_currency_code: String
) -> RawTransaction