Struct language_e2e_tests::account::AccountData

pub struct AccountData { /* fields omitted */ }

Represents an account along with initial state about it.

AccountData captures the initial state needed to create accounts for tests.

Methods

impl AccountData

pub fn new(balance: u64, sequence_number: u64) -> Self

Creates a new AccountData with a new account.

Most tests will want to use this constructor.

pub fn new_empty() -> Self

pub fn with_account(
account: Account,
balance: u64,
balance_currency_code: Identifier,
sequence_number: u64,
account_specifier: AccountTypeSpecifier
) -> Self

Creates a new AccountData with the provided account.

pub fn with_keypair(
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey,
balance: u64,
balance_currency_code: Identifier,
sequence_number: u64,
account_specifier: AccountTypeSpecifier
) -> Self

Creates a new AccountData with the provided account.

pub fn with_account_and_event_counts(
account: Account,
balance: u64,
balance_currency_code: Identifier,
sequence_number: u64,
sent_events_count: u64,
received_events_count: u64,
delegated_key_rotation_capability: bool,
delegated_withdrawal_capability: bool,
account_specifier: AccountTypeSpecifier,
is_frozen: bool
) -> Self

Creates a new AccountData with custom parameters.

pub fn add_balance_currency(&mut self, balance_currency_code: Identifier)

Adds the balance held by this account to the one represented as balance_currency_code

pub fn rotate_key(
&mut self,
privkey: Ed25519PrivateKey,
pubkey: Ed25519PublicKey
)

Changes the keys for this account to the provided ones.

pub fn sent_payment_event_type() -> FatStructType

pub fn received_payment_event_type() -> FatStructType

pub fn event_handle_type(ty: FatType) -> FatStructType

pub fn type_() -> FatStructType

Returns the (Move value) layout of the LibraAccount::T struct

pub fn account_type(&self) -> AccountTypeSpecifier

Returns whether the underlying account is an an empty account type or not.

pub fn to_value(&self) -> (Value, Vec<(Identifier, Value)>, Value, Value)

Creates and returns the top-level resources to be published under the account

pub fn make_account_access_path(&self) -> AccessPath

Returns the AccessPath that describes the Account resource instance.

Use this to retrieve or publish the Account blob.

pub fn make_balance_access_path(&self, code: Identifier) -> AccessPath

Returns the AccessPath that describes the Account balance resource instance.

Use this to retrieve or publish the Account blob.

pub fn make_event_generator_access_path(&self) -> AccessPath

Returns the AccessPath that describes the EventHandleGenerator resource instance.

Use this to retrieve or publish the EventHandleGenerator blob.

pub fn make_account_type_access_path(
&self,
account_specifier: AccountTypeSpecifier
) -> AccessPath

Returns the AccessPath that describes the Account balance resource instance.

Use this to retrieve or publish the Account blob.

pub fn to_writeset(&self) -> WriteSet

Creates a writeset that contains the account data and can be patched to the storage directly.

pub fn address(&self) -> &AccountAddress

Returns the address of the account. This is a hash of the public key the account was created with.

The address does not change if the account's keys are rotated.

pub fn account(&self) -> &Account

Returns the underlying Account instance.

pub fn into_account(self) -> Account

Converts this data into an Account instance.

pub fn balance(&self, currency_code: &IdentStr) -> u64

Returns the initial balance.

pub fn sequence_number(&self) -> u64

Returns the initial sequence number.

pub fn sent_events_key(&self) -> &[u8]

Returns the unique key for this sent events stream.

pub fn sent_events_count(&self) -> u64

Returns the initial sent events count.

pub fn received_events_key(&self) -> &[u8]

Returns the unique key for this received events stream.

pub fn received_events_count(&self) -> u64

Returns the initial received events count.

impl AccountData

pub fn strategy(
balance_strategy: impl Strategy<Value = u64>
) -> impl Strategy<Value = Self>

Returns a [Strategy] that creates AccountData instances.