Trait language_e2e_tests::account_universe::AUTransactionGen

pub trait AUTransactionGen: Debug {
    fn apply(
        &self, 
        universe: &mut AccountUniverse
    ) -> (SignedTransaction, (TransactionStatus, u64));

    fn arced(self) -> Arc<dyn AUTransactionGen>
    where
        Self: 'static + Sized,
    { ... }
}

Represents any sort of transaction that can be done in an account universe.

Required methods

fn apply(
&self,
universe: &mut AccountUniverse
) -> (SignedTransaction, (TransactionStatus, u64))

Applies this transaction onto the universe, updating balances within the universe as necessary. Returns a signed transaction that can be run on the VM and the expected values: the transaction status and the gas used.

Loading content...
Provided methods

fn arced(self) -> Arc<dyn AUTransactionGen> where
Self: 'static + Sized,

Creates an arced version of this transaction, suitable for dynamic dispatch.

Loading content...