Module language_e2e_tests::common_transactions

Support for encoding transactions for common situations.

Statics

CREATE_ACCOUNT_SCRIPT

EMPTY_SCRIPT

MULTI_AGENT_MINT_SCRIPT

MULTI_AGENT_SWAP_SCRIPT

Functions

create_account_txn
Returns a transaction to create a new account with the given arguments.

empty_txn

multi_agent_mint_script

multi_agent_mint_txn
Returns a transaction to mint coins from TC to DD to VASP.

multi_agent_p2p_script_function

multi_agent_p2p_txn
Returns a multi-agent p2p transaction.

multi_agent_swap_script

multi_agent_swap_txn
Returns a transaction to swap currencies between two accounts.

peer_to_peer_txn
Returns a transaction to transfer coin from one account to another (possibly new) one, with the given arguments.

raw_multi_agent_swap_txn
Returns an unsigned raw transaction to swap currencies between two accounts.

raw_rotate_key_txn
Returns a transaction to change the keys for the given account.

rotate_key_txn
Returns a transaction to change the keys for the given account.