Module language_e2e_tests::common_transactions

Support for encoding transactions for common situations.

Functions

add_validator_txn
Returns a transaction to add a new validator

create_account_txn
Returns a transaction to create a new account with the given arguments.

mint_txn
Returns a transaction to mint new funds with the given arguments.

peer_to_peer_txn
Returns a transaction to transfer coin from one account to another (possibly new) one, with the given arguments.

raw_rotate_key_txn
Returns a transaction to change the keys for the given account.

register_validator_txn
Returns a transaction to register the sender as a candidate validator

rotate_consensus_pubkey_txn
Returns a transaction to change the keys for the given account.

rotate_key_txn
Returns a transaction to change the keys for the given account.