Struct language_e2e_tests::executor::FakeExecutor

pub struct FakeExecutor { /* fields omitted */ }

Provides an environment to run a VM instance.

This struct is a mock in-memory implementation of the Libra executor.

Methods

impl FakeExecutor

pub fn from_genesis(write_set: &WriteSet) -> Self

Creates an executor from a genesis WriteSet.

pub fn from_genesis_file() -> Self

Creates an executor from the genesis file GENESIS_FILE_LOCATION

pub fn from_fresh_genesis() -> Self

Creates an executor using the standard genesis.

pub fn whitelist_genesis() -> Self

pub fn from_genesis_with_options(publishing_options: VMPublishingOption) -> Self

Creates an executor from the genesis file GENESIS_FILE_LOCATION with script/module publishing options given by publishing_options. These can only be either Open or CustomScript.

pub fn no_genesis() -> Self

Creates an executor in which no genesis state has been applied yet.

pub fn custom_genesis(
genesis_modules: Vec<VerifiedModule>,
validator_accounts: Option<usize>,
publishing_options: VMPublishingOption
) -> Self

Creates fresh genesis from the stdlib modules passed in.

pub fn create_accounts(
&mut self,
size: usize,
balance: u64,
seq_num: u64
) -> Vec<Account>

Creates a number of Account instances all with the same balance and sequence number, and publishes them to this executor's data store.

pub fn apply_write_set(&mut self, write_set: &WriteSet)

Applies a WriteSet to this executor's data store.

pub fn add_account_data(&mut self, account_data: &AccountData)

Adds an account to this executor's data store.

pub fn add_module(&mut self, module_id: &ModuleId, module: &CompiledModule)

Adds a module to this executor's data store.

Does not do any sort of verification on the module.

pub fn read_account_resource(
&self,
account: &Account
) -> Option<AccountResource>

Reads the resource [Value] for an account from this executor's data store.

pub fn read_balance_resource(
&self,
account: &Account
) -> Option<BalanceResource>

Reads the balance resource value for an account from this executor's data store.

pub fn read_account_info(
&self,
account: &Account
) -> Option<(AccountResource, BalanceResource)>

Reads the AccountResource and BalanceResource for this account. These are coupled together.

pub fn execute_block(
&self,
txn_block: Vec<SignedTransaction>
) -> Result<Vec<TransactionOutput>, VMStatus>

Executes the given block of transactions.

Typical tests will call this method and check that the output matches what was expected. However, this doesn't apply the results of successful transactions to the data store.

pub fn execute_and_apply(
&mut self,
transaction: SignedTransaction
) -> TransactionOutput

Executes the transaction as a singleton block and applies the resulting write set to the data store. Panics if execution fails

pub fn execute_transaction_block(
&self,
txn_block: Vec<Transaction>
) -> Result<Vec<TransactionOutput>, VMStatus>

pub fn execute_transaction(&self, txn: SignedTransaction) -> TransactionOutput

pub fn read_from_access_path(&self, path: &AccessPath) -> Option<Vec<u8>>

Get the blob for the associated AccessPath

pub fn verify_transaction(&self, txn: SignedTransaction) -> VMValidatorResult

Verifies the given transaction by running it through the VM verifier.

pub fn get_state_view(&self) -> &FakeDataStore

pub fn new_block(&mut self)